+++ LSTM.py
@@ -1,103 +1,3 @@
-
-
-import sys
-import numpy as npy
-import pandas as pds
-from sklearn.metrics import *
-import matplotlib.pyplot as plot
-from sklearn.preprocessing import *
-
-def squashingSigmoidFunc(xParameter):
-    sigmoidResult = 1 / (1 + npy.exp(-xParameter))
-    return sigmoidResult
-
-def squashingSigmoidDiff(yParameter):
-    sigmoidDiffResult = yParameter * (1 - yParameter)
-    return sigmoidDiffResult
-
-def tangentFunc(x):
-    tangentVal = npy.tanh(x)
-    return tangentVal
-
-def tangentFuncDiff(y):
-    tangentDiffVal = 1 - y * y
-    return tangentDiffVal
-
-def getArray(parameter):
-    arrayRes = npy.array(parameter)
-    return arrayRes
-
-def getSequences(df, winSize):
-    X = []
-    Y = []
-    len = df.shape[0]
-    itr =1;
-    while(itr <= len):
-        if itr + winSize + 1 > df.index[-1]:
-            break
-        locX = df.iloc[itr:itr+winSize]
-        locY = df.iloc[itr+winSize+1]
-        X.append(locX)
-        Y.append(locY)
-        arrayX = getArray(X)
-        arrayY = getArray(Y)
-        itr +=1
-    return arrayX, arrayY
-
-def split(X, Y, r):
-    len = X.shape[0]
-    splitData = int(len*r)
-    X1 = X[:splitData]
-    X2 = X[splitData:]
-    Y1 = Y[:splitData]
-    Y2 = Y[splitData:]
-    return X1, X2, Y1, Y2
-  
-def getRandomValue(frame, sdWeight, var,d):
-    frameX = frame[0]
-    frameY = frame[1]
-    if(d == 30):
-      val = npy.random.rand(frameX, frameY) * sdWeight + var
-    return val
-
-def getNumpyZerosLike(frame):
-    arr = npy.zeros_like(frame)
-    return arr
-  
-def getNumpyZeros(layr, val):
-    arr = npy.zeros((layr, val))
-    return arr
-
-def stackArray(vecA,vecB):
-    stack = npy.vstack((vecA, vecB))
-    return stack
-
-def getIpData(df):
-    IpData = df.iloc[:, [1]]
-    return IpData
-
-def readCSV(url):
-    file = pds.read_csv(url)
-    return file
-
-# Normalizing function
-sclr = MinMaxScaler()
-def fitScalerData(IpData):
-    sclr.fit(IpData);
-def transformData(IpData):
-    IpData = sclr.transform(IpData)
-    return IpData
-def downloadReadings(Readings):
-    Readings.to_csv('Readings.csv', index=False)
-    return True
-
-class SetParams:
-    def __init__(self, title, frame, sdWeight, var=0, d=30):
-        self.title = title
-        self.value = getRandomValue(frame,sdWeight,var,d)
-        self.diff = getNumpyZerosLike(self.value)
-        self.m = getNumpyZerosLike(self.value)
-
 class LongShortTermMemory:
     def __init__(self, LRate, maxItrs, stepTime, frame, hiddenLyrs=10, flag=True):
         self.LRate = LRate
@@ -234,118 +134,3 @@ class LongShortTermMemory:
           x+=1
         pred = npy.concatenate(predictY) 
         return pred
-
-inputURL = 'https://raw.githubusercontent.com/preerao7/FOREX-Rate-Prediction-using-RNN/refs/heads/main/dataset/EURUSD_D1_Sorted.csv'
-# Reading data from input file
-df = readCSV(inputURL)
-
-IpData = getIpData(df)
-
-  # Normalizing data
-fitScalerData(IpData)
-
-IpData = transformData(IpData)
-# Selecting target feature
-IpData = pds.DataFrame(IpData)
-
-LRateList = [0.01, 0.1]
-# LRateList = [0.1]
-maxItrsList = [50, 100,150,200]
-# maxItrsList = [50, 100]
-hiddenLyrsList = [1, 2, 3]
-stepTimeList = [2, 4]
-
-
-Readings = {
-    'LRate': [],
-    'hiddenLyrs': [],
-    'stepTime': [],
-    'maxItrs': [],
-    'r2Score': [],
-    'MeanSquaredError': [],
-    'RootMeanSquaredError': []
-}
-
-for hiddenLyrs in hiddenLyrsList:
-    for LRate in LRateList:
-        for stepTime in stepTimeList:
-            seqs = getSequences(IpData, stepTime)
-            xSeq = seqs[0]
-            ySeq = seqs[1]
-            xTrain, xTest, yTrain, yTest = split(xSeq, ySeq, 0.7)
-            for maxItrs in maxItrsList:
-                print('------------------------------------------------------------------------------------------------------------')
-                model = LongShortTermMemory(LRate, maxItrs, stepTime, frame=(1, 1), hiddenLyrs= hiddenLyrs)
-                print('No. of Hidden Layers:','\t',(hiddenLyrs))
-                print('Rate of Learning:','\t',(LRate))
-                print('Step Time Interval:','\t',(stepTime))
-                print('No. of Maximum Iterations:','\t',(maxItrs))
-                print('------------------------------------------------------------------------------------------------------------')
-                
-                model.modelFitting(xTrain, yTrain)
-                yPrediction = model.Prediction(xTest)
-                
-                r2Score = r2_score(yTest, yPrediction)
-                MeanSquaredError = mean_squared_error(yTest, yPrediction, squared=True)
-                RootMeanSquaredError = mean_squared_error(yTest, yPrediction, squared=False)
-
-                Readings['LRate'].append(LRate)
-                Readings['hiddenLyrs'].append(hiddenLyrs)
-                Readings['stepTime'].append(stepTime)
-                Readings['maxItrs'].append(maxItrs)
-                Readings['r2Score'].append(r2Score)
-                Readings['MeanSquaredError'].append(MeanSquaredError)
-                Readings['RootMeanSquaredError'].append(RootMeanSquaredError)
-
-Readings = pds.DataFrame.from_dict(Readings)
-confirm = downloadReadings(Readings)
-if(confirm):
-    print("Readings.csv has been downloaded!")
-
-def loopHiddenLyrs(hiddenLyrs):
-    def loopLRates(LrateList):
-        def loopStepTime(stepTimeList):
-           for stepTime in stepTimeList:
-              
-              y = Readings.loc[(Readings['hiddenLyrs'] == hiddenLyrs) &
-                                    (Readings['LRate'] == LRate) &
-                                    (Readings['stepTime'] == stepTime), [mx]]
-              ax[i, j].plot(maxItrsList, npy.array(y))
-              Title = 'Layers_Hidden: '+ str(hiddenLyrs) + ' Learning_Rate: ' + str(LRate)
-              ax[i, j].set(xlabel='Iterations', ylabel=mx,
-                            title=Title)
-              ax[i, j].legend(['Time step - {}'.format(stepTime) for stepTime in stepTimeList])
-        for j, LRate in enumerate(LRateList):
-            loopStepTime(stepTimeList=stepTimeList)
-    for i, hiddenLyrs in enumerate(hiddenLyrsList):
-        loopLRates(LrateList=LRateList)
-
-for mx in ['r2Score', 'MeanSquaredError', 'RootMeanSquaredError']:
-    fig, ax = plot.subplots(2, 2, figsize=(11,11))
-    loopHiddenLyrs(hiddenLyrs)
-    plot.show()
-
-#Best Hyper Parameters Choosen after Result Obervations
-LRate=0.01
-maxItrs=100
-stepTime=2
-hiddenLyrs=5
-
-model = LongShortTermMemory(LRate, maxItrs, stepTime, frame=(1, 1), hiddenLyrs= hiddenLyrs)
-model.modelFitting(xTrain, yTrain)
-
-yPrediction = model.Prediction(xTest)
-
-r2Score = r2_score(yTest, yPrediction)
-MeanSquaredError = mean_squared_error(yTest, yPrediction, squared=True)
-RootMeanSquaredError = mean_squared_error(yTest, yPrediction, squared=False)
-
-print('R2 Score:\t',r2Score)
-print('Mean Squared Error "MSE":\t',MeanSquaredError)
-print('Root Mean Squared Error "RMSE":\t',RootMeanSquaredError)
-
-# Plot Prediction Graph
-plot.plot(yTest, label="Actual")
-plot.plot(yPrediction, label="Prediction")
-plot.legend()
-plot.show()
